requirements.txt
streamlit==1.46.0
openai==1.93.0
google-genai==1.23.0
python-dotenv==1.1.1
anthropic==0.55.0